fix: resolve all clippy warnings
- Remove empty line after doc comment

- Fix doc list item indentation

- Remove unnecessary unsafe blocks in nested contexts

- Simplify identical if-else branches

- Use vec![] macro instead of Vec::new() + push()

- Add #[allow(dead_code)] for intentionally unused helper methods

+++ src/tag_path.rs
@@ -283,7 +283,7 @@ impl TagPath {
                 // Add array index
                 path.push(0x28); // Element segment
                 path.push(0x04); // Size: 4 bytes for 32-bit index (DINT)
-                let index_u32 = *index as u32;
+                let index_u32 = *index;
                 path.extend_from_slice(&index_u32.to_le_bytes());
             }
         }
